% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_dataframe_prime.R
\name{flatten_dataframe_prime}
\alias{flatten_dataframe_prime}
\title{Flatten Dataframe Prime}
\usage{
flatten_dataframe_prime(df)
}
\arguments{
\item{df}{A dataframe of numbers; largest number cannot be greater than 1000}
}
\value{
A numeric vector that has all of the prime numbers in df
}
\description{
Flattens a list so that all prime numbers in embedded lists are 
returned in a flat list
}
\examples{
a <- data.frame(c(2, 3, 4), c(5, 6, 7), c(8, 9, 10))
flatten_dataframe_prime(a) 
# returns the numeric vector c(2, 3, 5, 7)

}
